Add service classes for book, file, loan, and user management
